Update patch set 1
Patch Set 1:

Here's the scenario running through my head. If a user upgrades their aboot (right, they shouldn't) and tries to flash a build that does not have that bootloader in valid_bootloaders yet, its going to flash a virgin boot.img to it. This will fail to boot 100% of the time.

When dev devices come out, we could make another file, call it dev_bootloaders, that checks for a dev device and flashes this virgin boot.img to it.

If history repeats, we should be able to tell the difference. On d2vzw, the same iteration of the dev and regular device had different release versions:
* OYALG1 <-- dev device
* VRALG1 <-- regular device

I think not flashing a virgin boot.img to a device is a better option than flashing it knowing that it will never work.

Patch-set: 1