Update patch set 1
Patch Set 1:

The asserts is this script won't stop the rom from being installed, they are just controlling when a device gets a patched boot and when not. Development bootloaders won't be an issue unless they have the exact same build numbers the locked ones do, which is highly unlikely.  At that point if development bootloaders carry the same variant names, we may have to add back an else and alternate unpatched boot flash. The whole point is to differentiate between known locked bootloaders and everything else, and patch boot on devices with the known locked bootloaders. Everything else gets an unpatched boot image. 

Yes, recoveries for att/vzw must be patched.  That's not a big deal, it's just different.  

This looks right/clean.  if the paths are all correct then this will work just as it worked in testing before all the cleanup.

Patch-set: 1