Update patch set 1
Patch Set 1: Code-Review-1

Requires the new blob, located at the usual place.  Also, it appears that the only difference between the new blob and the existing one is the addition of the setPrivacy, which is interesting.

Now, the real issue is that we really shouldn't use this on devices that don't use the privacy lock.  Even if the baseband supports it, if stock ROM doesn't support/use the privacy lock, then the opposite problem can occur--that is, users who flash back to stock may find the GPS won't work there with no way to resolve it.

So, the real question is, which devices are affected?  I know jflte{spr,vzw} are.  Those affected should have the symbol _ZN16LocApiV02Adapter10setPrivacyEa in stock ROM /system/lib/libloc_api_v02.so.

Is it all CDMA, no GSM?

If that's the case, I'll add a system property check to see if it's a CDMA device, and only attempt set the privacy lock in that case.

Patch-set: 1