Update patch set 2
Patch Set 2:

So it appears that none of the GSM devices support the privacy lock in TW.  For PS #2, I've made it so we always enable the lock when turning on the GPS, but never disable it again.  This way we can make sure the GPS works on any device if the user is coming from TW, and we can also ensure that the GPS will continue to work should the user flash back to TW.

Can someone test PS #2 on a GSM device to make sure there's nothing wonky with the baseband such that issuing the superfluous setPrivacy command doesn't break anything?

Patch-set: 2