Update patch set 1
Patch Set 1:

They are also doing some management of a cgroup which controls which processes will use compcache.

It makes a lot of sense actually.. You don't want foreground stuff using it, because some additional latency is introduced and the UI can lag. It's great for background services and hidden apps, though.

Patch-set: 1