Update patch set 1
Patch Set 1:

Hmmph.  I'm still not able to repro following your exact sequence of events (jflteatt).   Will wait to see what cyanogen finds.

On the other issue - no notification sounds: that's not resolved by reverting this, right ?  Assuming this is the case, any idea whether this has always been present in cm-10.2 and/or whether it's present in cm-10.1 ?

Patch-set: 1