Update patch set 1
Patch Set 1:

@Steve, I have been able to reproduce the no notifications/Alarm playing problem and I have a suspicion what the cause might be...

I have my phone overnight in the Samsung Desktop Dock to charge.   If I connect a BT headset to it while its in the dock (or insert it in the dock after the BT headset is conected), listen to music and then turn off the BT headset/disconnect, subsequently the Alarm will not play through the phone speaker until rebooted.  

I suspect it is trying to route the audio via the usb port to the dock (which it thinks has a speaker connected?)

Could it be the usb audio support introduced the regression?

I am still trying to figure out the exact steps to reproduce the in-call mic issue reported above

Patch-set: 1