Update patch set 1
Patch Set 1: Code-Review+1

Making a sequence of voip calls and then finding non-voip calls have no audio is true irrespective of the use of a headset.  ie it's equally broken on handset too.

This being the case, I see no real blocker for merging this.  Back to +1.

Back to the very above - there are two problem case scenarios I've found so far:

1) Make several voip calls in a row (I'm testing with localphone).  Now make a non-voip call and I usually find I have mute tx/rx audio.

2) Dial your own non-voip number from your voip number.  results in mute tx/rx audio after a single call every time.  Resolvable by overwriting the presence of a voip related verb at the head of audio-caf enableVoiceCall.  This doesn't feel like the right root cause fix but it's starting point.

The workaround for 2) does not help with 1).  (Although I've not yet tried also overwriting play voip mod in the same place.)

For reference, I'm jflteatt with MF3 modem and MDL non-hlos.bin

My time is limited so if anyone else is able to investigate too, great.

Patch-set: 1
Label: Code-Review=+1