Update patch set 1
Patch Set 1:

sensors are fine and gps is fine and it is possible that the google edition may not have the capibility of the regular edition since they rebase the code onto caf and didn't opimize it for battery life because they wanted the  4.3 out fast.  try 4.2 modem. it should work far better than 4.3 due to samsung already optimize it.

Patch-set: 1
Label: Code-Review=0