Update patch set 1
Patch Set 1: Code-Review+1

Chris: This does work when the sensors are working :) However, for the sensors to work, I had to revert sensorservice to the Android 4.3 version. It's probably better to wait for the Kit Kat google edition blobs instead of trying to work around that issue now though.

Patch-set: 1