Update patch set 1
Patch Set 1:

this probably needs more work - like the rest of the original commit minus whatever partition is borked.  i dont know which one it is yet.

i cant logcat when it hangs, so im just moving in parts of the original patch a little at a time to troubleshoot, hopefully somebody has a better way.

Patch-set: 1