Update patch set 24
Patch Set 24:

I lied and it hasnt been fixed on github yet (thought it was the tip, clearly not...) I'll be fixed shortly, or I'll submit a revert or something.

As far as recovery goes, check out releasetools/loki.sh
The zip will do it by itself if needed.

As far as including it in recovery, you'd need to have recovery check the cmdline and determine whether or not its necessary for that particular device.

Patch-set: 24