Update patch set 1
Patch Set 1:

was just curious as the reasoning was not really true in that point.

though i havent experienced any changes with it as well and see it too as not needed.

but now its the question, reverting and adding another pointless commit to the device tree, making it more messy or just ignoring it, as removing it and/or keeping it does not change anything?

Patch-set: 1