Update patch set 2
Patch Set 2:

at the end of the day jf init is all a mess and im working on merging ml4, and this difference just annoyed me.  
we could also just restore it in the 'bigger update' but the revert ties it to an explanation and stops annoying me until then.

not a huge deal.

Patch-set: 2
Reviewer: dcd <7874@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0