Update patch set 3
Patch Set 3:

Alexander: there's 2 issues here, at least on d2, I believe Dan has copied our implementation on d2. First when the phone is plugged in and booting, the light changes from pulsing to charging, and then you write 0, which turns the led off even if it was charging. 

Second is charger mode. I agree changing it to bootanim solves that problem but what about the other?

Patch-set: 3