Update patch set 3
Patch Set 3:

yes, on jf too as the led light overrides the pulsing.

it would need to be delayed but at least with the comments here it works if plugged in and plugged out, as well as if on offline charging mode.

only "ugly" effect is the led overriding the pulsing, as you said.

Patch-set: 3