Update patch set 1
Patch Set 1:

fyi: i want to de-unify jflte

I might not get to it for a week or two though, I dont have my jfltevzw with me at the moment.

I've already done it for d2vzw and d2spr on the cm-12.0 branches, so if you want to jump ahead and go for it, check those out.

Either way, there's still much commonized bringup that needs to be done. Much of the lollipop bringup work will be commonized anyway.

Patch-set: 1