Update patch set 2
Patch Set 2:

It is designed to be vendor neutral. It is an internal API meaning it can be used for the opt/telephony/internal and that all devices can use. Discuss with others so we can use that API in telephony, not in device part. device part is in RIL.java.

Patch-set: 2