Update patch set 2
Patch Set 2:

doc - an LCD is most certainly going to beat up its battery by turning on like this all the time. The benefit an AMOLED has is that true black pixels do not take power.

I'm not saying it doesn't work to show the notifications on flo/m7, but since they are both LCD panels, and don't have the doze stuff that bacon does, the panel is going to suck down just as much power as if it was on normally.

Patch-set: 2