Update patch set 1
Patch Set 1:

So, updater-script makes the assumption that your partition is whatever is in your fstab, and uses that to mount.

mount("ext4", "EMMC", "/dev/block/platform/msm_sdcc.1/by-name/userdata", "/data", "");

If there are multiple instances of the same partition, the second one appears to be used.

That coupled with recovery likely having issues mounting partitions, this stuff just cant go in.

Plus, we can't ship busybox and blkid like this. To even be remotely considered, it would have to be cleaned up a ton.

Patch-set: 1