Update patch set 1
Patch Set 1:

i understand that this wont be merged, but few questions regardless, since youre going to pick it anyways for your builds, and maybe i will.

are you basing this on the latest fstab?  we changed boot/recovery flags to 'recoveryonly' and removed length=-16384 from data a while back, so i cant figure out if youre working off of an old cm-10.2 fstab or something?

also, why is the duplicate copy of fstab necessary?

i would think you could just sed the existing fstab in place (and make no modifications to fstab in this change) and accomplish the same thing.

Patch-set: 1