Update patch set 1
Patch Set 1:

When 20150112 or 20150113 nightly is installed, filesystem size of /system becomes 600MB, while the partition size is still 2.8GB on i9505.

This makes the /system filesystem too small for practical use.

I'm not sure how recovery works, but AFAIK there are at least two different methods two install - you either untar contents to the filesystem (possibly preceeded by mkfs) or you flash binary image to the block device - so either that's happening and the image is wrong size, or the filesystem is created with wrong size.

Patch-set: 1