Update patch set 1
Patch Set 1:

Because the google play edition has a smaller system partition. The builds wont even flash on it.

The info I had received implied that this was the correct size, but it turns out its likely double this.

I have another commit up to rectify this: http://review.cyanogenmod.org/#/c/85165/

But I'm waiting on confirmation that it works on a google play edition.

Patch-set: 1