Update patch set 2
Patch Set 2:

@DCD

I probably should have tried building with this reverted before posting in it's post....because after I did it had no affect on the issue.  Hindsight is 20/20


But to clarify the issue:
Starting around the time this was merged 1/15 - 1/18.  My jfltevzw has been unable to read the external SDcard aka SDCARD1.


The phone sees it mounted vs unmounted, but cannot access files.  Does not appear in Settings/storage.  Root apps, cannot access the info on the card (ie Root explorer). Google Music cannot read songs on externalSD.


Logcats without obvious error, and dmesg and kloripper without audits referencing the external partition.


So perhaps, the better question is, 'Is anyone else having this behavior?' 


And Yes the SDcard is fine.

Patch-set: 2