Update patch set 1
Patch Set 1:

MMI is MultiMedia Interface for my car.  

Without this, it works fine on 'dumb' bluetooth head sets and speakers.  But will not pair to the car.

I am not aware of any hardware BT differences in the jflte models, perhaps some of the caf merged into BT and bluedroid over the last week may explain the change in behavior.

If someone wants me to run off jflte*** build for them to test.... just let me know.

But for the first time in a couple weeks, my BT is fully operational.

Patch-set: 1