Update patch set 1
Patch Set 1:

Ok, another reason was to keep the symmetry between the device and the vendor directories. But, if you say it doesn't matter, I'll abandon the commit and add the missing part at least in jfltexx. Some problems may raise when extracting from multiple jf devices. Are you sure about this?

Patch-set: 1