Update patch set 2
Patch Set 2:

> Could you possibly grab a log?

http://pastebin.com/j0E81nVa

Connected to my bluetooth headset and called myself.
Let it ring for a bit, saw a Pushbullet notification that I was getting a call, but nothing showed that on the phone except for the LTE -> HSPA+ network change.
Let it ring for a while before turning off my headset and turning off Bluetooth, then my phone showed the incoming call.
Declined the call on my phone then stopped the logcat.

Patch-set: 2