Update patch set 2
Patch Set 2:

We are using the following for d2, is this valid?
on property:sys.boot_completed=1
    write /sys/block/mmcblk0/queue/scheduler bfq
    write /sys/block/mmcblk1/queue/scheduler bfq
    write /sys/block/mmcblk0/queue/iosched/slice_idle 0
    write /sys/block/mmcblk1/queue/iosched/slice_idle 0

We are doing this for each block mmcblk0,mmcblk1. Do we have to do this for all the block or only 0?

Patch-set: 2
Reviewer: Matrix Zone <12095@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0