Update patch set 26
Patch Set 26:

This fixes issues with bluetooth, and some other issues related to the legacy HAL and M. The new HAL is also much cleaner and easier to maintain, and will make the transition to N much easier. It's very likely that the legacy HAL won't be carried over to the next CM version. Thus without this, no working audio.

Patch-set: 26