Update patch set 2
Patch Set 2:

And I think the benefits of the franco thermal driver are just #1 that we aren't routing a bunch of thermal calls through userspace, which is inefficient, but rather we are handling it all through the kernel. Also, #2 we never have to worry about thermal-engine being incompatible with newer versions of android, rather we can rely on always having thermal controls.

Patch-set: 2