Update patch set 35
Patch Set 35:

Grando: Audio is completely broken only for headphones without a mic, which is weird. I'll get a log of it tomorrow if I remember, or maybe later today if I can.
David: Huh, weird. Is it no in call audio at all, or only when speakerphone, or what/when? More detail would be appreciated.

Patch-set: 35