Update patch set 6
Patch Set 6:

> Added acdbid for SND_DEVICE_IN_HEADSET_MIC_FLUENCE

Hey Arne I'm wondering if this requires some other changes. My line of thinking is that fluence is only used for internal mics so in case the headset doesn't have a mic it should use output of headset and input of voice-speaker-mic if I'm not mistaken, but in the hal this usecase also points to "headset-mic". So I'm not sure if my understanding is correct or if we need to do something else to ensure it uses the speaker mic and not try to use the non-existent headset mic?

Patch-set: 6