Update patch set 35
Patch Set 35:

Paul, Here's more information.

I've factory reset my phone.. JFLTEVZW

Can confirm the following:

Sound is working just fine..
Sound recorder is working just fine..
Sound record playback is fine.. tested by this is rubbing the microphone on the bottom of the phoe.

However,
- Using the vanilla android dialer..
- I'm able to answer calls and place calls..  
- i can hear touch tones go out while on call (while on speakerphone and not on speaker phone) and confirmed this on the other line, so that's working..

But..  
While on call while not on speaker phone
- no audio in (cant hear anything)
- no audio out (nobody can hear me except for touch tones)

While on call on speakerphone 
- no audio in (cant hear anything)
- no audio out (nobody can hear me except for touch tones)

I am unable to test the headset to see if this happens on the headset too while plugged in.

Patch-set: 35