Update patch set 7
Patch Set 7:

> You can try these:
 > 
 > https://drive.google.com/open?id=0B4Gcy6G7cgo8TWItcVZwaGd1Smc
 > https://drive.google.com/open?id=0B4Gcy6G7cgo8OUdOS3l2a210MXM
 > 
 > I think we can remove SND_DEVICE_IN_HEADSET_MIC_FLUENCE. Doesn't
 > look relevant for us.

I'll give those a shot. As far as the rest of the voice calling I got it working using these files mostly want to focus on the voice-dmic-ef and voice-speaker-dmic-ef they are the same as dmic-endfire and speaker-dmic-endfire except the internal route config. In the snd_soc Call main mic was the same as dmic-endfire, except the acdbid and the internal route config. These files have lots of extra stuff, because I never bothered removing, but mostly to get calling working you'll just need the dmic entries. 

http://textuploader.com/5igdw

http://textuploader.com/5igdy

Going to see if changing the acdbloader works or if something else needs to be done for headphones.

Patch-set: 7