Update patch set 7
Patch Set 7:

I've changed our mixers to match the way the other samsung phones are doing using this is a base and its working great. I still wanna test it AEC and NS actually works. I'll have to test that later tho. I saw the change to Telphony to enable it, but does the our hal support it? Eitherway I'll have to give it a try sometime during the week. I can put up the modified mixer_paths and audio_platform_info if you are interested.

Patch-set: 7