Update patch set 12
Patch Set 12:

Works fine with the old hal. But looks like there was an old patch on there also to make 3-pole headphones work. http://review.cyanogenmod.org/#/c/44791/ there is also a patch on the same date for the jf-common directory that coincides with it. I'm going to see if I can figure out how to port it over for the new HAL.

Patch-set: 12