Update patch set 12
Patch Set 12:

The issue we are having is that when you use a regular headset w/ mic it works great no errors since it routes voice-call audio to the ear pieces and the voice-call input is taken from the headset mic. The issue comes when we try to use headsets w/o mic the hal tries to use headset for voice-call audio out and voice-speaker input, but for our device it throws up an error of it being a "invalid device pair". For our device the voice-call audio out is sent to headphones and the input device should be the voice-main-mic instead of speaker.

Patch-set: 12