Update patch set 14
Patch Set 14:

Well I looked at libaudcal.so really quickly it seems to export acdbcmdgetdevicepair or something similar. Don't have that information with me. If someone can take a look and figure out a way that we can let voice-speaker and headphone not throw out the invalid device pair error it would be great, because then we don't need to mess with the hal.

Patch-set: 14