Update patch set 24
Patch Set 24:

You know, the thing is, headphones w/o mic only worked on one ear on the Old HAL as well. If that's the last remaining issue, is it worth it to spend the time to fix it? And if we couldn't fix it on the last HAL, how difficult would it be to fix it in the new one?
Also, have audio (and mic) usecases not regarding to calls been tested? (e.g. Music, Audio Recording, etc.)

Patch-set: 24