Update patch set 24
Patch Set 24:

If that's the only issue, and it seems that it is, I think it's time for it to be merged imo. If Zhao and/or Dan could take a look at this commit to make sure it all looks OK (and I'm running a build with it at the moment to test it myself), it's probably time to merge it. If we don't lose any functionality from the Old HAL, we can try to fix it in the future and leave it as it is now for nightlies.

Patch-set: 24