Update patch set 6
Patch Set 6:

> Uploaded patch set 6.

So I'm not sure if this is intended, but I've applied all the pending changes for jf and multicore powersaving and temperature throttling seem to be disabled by default or at least that is what it shows in Kernel Auditor. There were also two instances while chariging my phone at night it was at 90% in the morning and it would just drain even if it was on the charger. I thought it might be the bionic change that caches the host file so built without it today so going to see how it goes. Couldn't grab logs tho. I can try to if it happens again.

Patch-set: 6