Update patch set 4
Patch Set 4:

It probably performs like shit on anything slower than an 8974. We turned it off for all our 8916s. The reason is that this is all done on the GPU with a shader since MDP doesn't do blurring. There is a patch to speed it up in the 8974 display HAL (patch allows for mixed-mode composition) but I don't think it would be applicable to any prior hardware.

Patch-set: 4