Update patch set 6
Patch Set 6:

You are probably facing crashes.  This commit came about because slayer increased the size from 1K to 2K to fix JIRA CYAN-5239.  I didn't want to deal with this again the future so I made it dynamic instead.  But the 1K buffer was definitely not big enough for all remotes out there.

Patch-set: 6