Update patch set 4
Patch Set 4:

> Uploaded patch set 4.

I'm not sure if you plan on doing the sepolicy stuff, but was wondering if you could do the sdcardfs stuff too if you do. I've been running it on cm13 with it set to permissive and everything seems to be working on permissive. Just don't have any free time right now to fix it so I put it up hoping someone else might be able to take a look into fixing the denials.

Patch-set: 4