Update patch set 2
Patch Set 2:

Oh, I also remember, Samsung supported this prop in 4.2 blobs, but not after 4.4. If you search the blobs for it, I'm pretty sure you won't find it. Even if they set it in build.prop, it's actually just a no-op, since the ril blobs aren't aware of it.

Patch-set: 2