Update patch set 1
Patch Set 1:

I don't think there is significant difference in first boot time, unless I needed to clean build for this change to take place. Waiting for a minute and a bit for it to boot isn't that bad, I guess. I'll abandon it, and find a more useful thing to change!

Patch-set: 1