Update patch set 3
Patch Set 3:

> *Remove logs.no reason to even exist
Why would we truncate logging functions? Does this wrapper not output logs? If it does, we need to leave it intact. If it does output logs, we should ifdef it so that in ENG builds we get logging.

> *remove unnecessary code
I'll need to take a closer look at the extent of what was removed here soon.

Patch-set: 3