Update patch set 2
Patch Set 2: Verified+1

So, in testing, I found that while playing "Or Nah" by Ty Dolla Sign as test audio, playing from the S4 at maximum media volume, recording Db output on both a Nexus 6 and Pixel XL 2, the data is as follows:

On Stock (5.0.1 VRSGPL1, which doesn't use a mixer_paths.xml configuration), we recorded an average output of 46 dB, and max of 59 dB. I also noted that "s" consonant's sounded sharp in comparison with other devices.

On the 11/08 build (using a mixex_paths.xml audio configuration with the speakers set to 78 as the output value), we got an average output of 44 dB and a max output of 57 dB.

Upon @Side's reccomendation, I tried the value of 80 for the speaker output in mixer_paths.xml, and with that we got an average output of 46 dB, and a max output of 58 dB (notably closer to stock on the average value). 

I then tried 82, and with that we got an average output of 49 dB, and a max output of 59 dB (notably closer to stock on the maximum output). 

Splitting the difference at 81, we got an average output of 46 dB, and a max output of 59 dB (identical to stock, though notably, the average was higher values of 46, i.e. hung around 46.8/9, while stock hung around 46.5, but I'll call this close enough). 

Tl;DR: mixer_paths value 81 matches stock's dB output within a few tenths of a dB (as close as we'll get in my opinion).

Patch-set: 2
Reviewer: Nolen Johnson <16185@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1