Update patch set 2
Patch Set 2:

> jfltespr has been broken since at least march apparently, so i'd
 > guess vzw is too.  is anybody actually using cdma jf anymore?
 > 
 > anyways, i started down this path and fixed it somewhere along the
 > way and didn't really feel like tracing backwards.  this change as
 > is does bring jfltespr back to life.

I am, and though I haven't tested RIL in a few weeks, it worked as soon ago as August, and I have users actively using VZW as their DD with no complaints... Maybe just SPR is broken?

If it is, we need to figure this out in a different way than this, as the amount of testing we'd need to go through carrier by carrier is something I don't want to do until O.

As for RIL being broken on SPR, it isn't on the build roster in part b/c it isn't maintained... are  you just finding it to be broken in your personal builds/testing?

Patch-set: 2