Update patch set 2
Patch Set 2:

> yeah I have no desire to maintain jf again. someone handed me
 > jfltespr and said radio doesn't work so I'm handing it back
 > working.
 > 
 > I suspect the problem is dataservices, if I get time this weekend
 > I'll try that with old ril before I hand off the device.

Ahhh. I see. Can you test if this RIL works on SPR? Talking Outgoing and Incoming calls, as well as in call audio, etc. Then I can test on VZW, and we can then test on jfltexx, and if it works on xx, we can assume its good for GSM carriers.

If it truly works, I am not opposed to merging this, I am just noting that I don't want to do significant testing across variants lol.

Patch-set: 2