Update patch set 2
Patch Set 2:

LTE works for me but I'm unable to test voice on the account. I can try a GSM sim in it but I doubt you're interested in that use case. I only have 48 more hours with the device. So I can't really commit to testing, and I follow your reasoning.

dataservices is probably the problem.

Patch-set: 2