Add missing SELinux policy for proximity sensor calibration on Samsung.
Fixes some Samsung phones (notably the i9505 but there are similar reports)
from getting a blank screen and apparent "hang" when making/receiving calls.

This is due to these missing policies in SEAndroid on these phoens which
prevent the proximity sensor from having it's calibration reloaded after
each activation by system_server, which means even a calibration app
can only fix the problem once before the defaults are re-asserted.

This problem only manifests if high-low thresholds don't match your
proximity sensor as supplied, and does still require a calibration to
fix, but allows any of the calibration apps on the Play store to correctly
do it.

Change-Id: Id67689faad289dabae8e48310e4ac6f9b34acc41