Update patch set 4
Patch Set 4:

> i would agree with Nolen.Oreo comes with built-in BurnIn feature

You have the right to agree, but I am not even disagreeing. What I am saying is that the current patch does nothing because this custom feature that we had in 14.1 was not ported to 15.1.

Patch-set: 4
Reviewer: Bruno Martins <13648@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>