Update patch set 2
Patch Set 2: Verified+1 Code-Review+2

Changed author to the author of the mkbootimg, and updated commit msg. Thanks for the contributions.

Way back when I was waiting for the hardware/samsung/exynos split that was purported to be happening. With that being dead, this is good to go.

Patch-set: 2
Reviewer: Nolen Johnson <16185@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=+2