Update patch set 1
Patch Set 1:

Looks like the commit message I wrote got lost during rebasing, but basically the
reason is because we need to run the service as root due to the qcom variant having
to use the wifi driver as kernel module. Same as the situation with serrano.
I might also have to switch the other variants to module as well, due to issues
with oversized recovery image.

Patch-set: 1