Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> as far as I can see, the only thing we do here is drop
> 
> -    <WizardAction wizard:uri="intent:#Intent;action=org.lineageos.setupwizard.LINEAGE_DATETIME;end" id="datetime">
> 
> Why do we need to do that?

Because it's broken and causes issues with date/time. At least that was the case several android versions ago, when I reported it. Nothing was done about it though. If I remember correctly the handling of it was rewritten, and it was fine before that. It's also kind of useless on a phone, which will get the date/time from the mobile network, and if not it will get it from the wifi network.

Patch-set: 1