fixup! jf-common: init: Add SGH-M919V
According to some other sources, this model was initially designed for
Videotron devices.

For whatever reason, it does support LTE on that carrier. It is unknown
why it is thought to be otherwise on other carriers, though the theory
is that it supports select different frequency bands that were initially
not used widely, but became a non-issue towards the end of the device's
life cycle as carriers started adopting more frequency bands.

Update the incorrect code comment.

Change-Id: Ibe782f3b66bf88c2a092b767187cf9e898c32c26